+++ ufo.h
@@ -20,7 +20,7 @@ namespace ufo
     struct file {
         string name;
         string path;
-        int size;
+        unsigned long size;
         vector<int> dateCreated;
         vector<int> dateModified;
     };
